redis-service/requirements.txt
Flask==3.0.3
redis==5.0.4
gunicorn==21.2.0
PyJWT==2.9.0
cryptography==42.0.8
requests==2.32.3
flask-cors==4.0.0
